Make render pipeline stages easier to maintain
The renderer pipeline had accumulated data-fetching, page-data, CSS, and cache-key helpers in one large class. This keeps RenderPipeline as the orchestration shell and moves stage logic into focused modules while preserving the private CSS hook used by behavior tests.

Constraint: Task required a narrow split of src/rendering/orchestrator/pipeline.ts into stage modules

Rejected: Rewrite renderPage itself | too much behavior risk for a refactor lane

Confidence: high

Scope-risk: moderate

Tested: deno check modified pipeline files

Tested: deno lint modified pipeline files

Tested: focused pipeline/module/css-cache tests

Tested: deno task typecheck
